[#13] 🩹 회원 등록 카드 가격 총합 속성 타입 수정

+++ domain/src/main/kotlin/com/splab/springgames/domain/member/domain/Member.kt
@@ -33,7 +33,7 @@ data class Member(
                 registeredDate = registeredDate,
                 level = Level.BROZNE,
                 gameCardTotalCount = GameCardTotalCount(0),
-                gameCardTotalPrice = GameCardTotalPrice(0)
+                gameCardTotalPrice = GameCardTotalPrice(0.toBigDecimal())
             )
         }
